+++ frontend/src/components/Introduction/Introduction.tsx
